bug fixing

+++ reversi/ai/ai.hpp
@@ -31,7 +31,7 @@ namespace reversi
 
 		int win_tile;
 
-		int eval(const state &cstate);
+		float eval(const state &cstate);
 		int pos(const std::array<int, REVERSI_BOARD_SIZE> &board,
 			int x, int y);
 		void print_board(const state &cstate);
@@ -41,8 +41,8 @@ namespace reversi
 		bool check_board(state &test, int x, int y, int tile);
 		std::vector<state> get_next_states(const state &cstate,
 						   int tile);
-		int minimaxab(const state &cstate, int depth, int alpha,
-			      int beta, bool max, int tile);
+		float minimaxab(const state &cstate, int depth, float alpha,
+			      float beta, bool max, int tile);
 
 		// Re-entrant version of minimaxab
 		boost::optional<state>  minimaxab_r();
